feat: Implement Search Completer for location suggestions
- Added SearchCompleter class to handle location autocomplete using MKLocalSearchCompleter.
- Included distance calculation from user's location for search results.
- Created a new SettingsView for user settings management.
- Added SkeletonCarouselCard for loading placeholders in UI.
- Updated LoginView to replace glassEffect with cornerRadius for image styling.
- Enhanced image upload functionality in backend to include image resizing and compression using sharp.
