+++ Sidequest3/Karte.swift
@@ -1,64 +1,14 @@
 import SwiftUI
 import MapKit
-import CoreLocation
-import Combine
-
-class LocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
-
-    private let manager = CLLocationManager()
-    private var hasSetInitialPosition = false
-    var positionOverridden = false
-    @Published var lastLocation: CLLocation?
-    @Published var position: MapCameraPosition = .region(MKCoordinateRegion(
-        center: CLLocationCoordinate2D(latitude: 52.5200, longitude: 13.4050),
-        span: MKCoordinateSpan(latitudeDelta: 0.05, longitudeDelta: 0.05)
-    ))
-
-    override init() {
-        super.init()
-        manager.delegate = self
-        manager.desiredAccuracy = kCLLocationAccuracyBest
-        manager.requestWhenInUseAuthorization()
-        manager.startUpdatingLocation()
-    }
-
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        guard let location = locations.first, !hasSetInitialPosition else { return }
-        hasSetInitialPosition = true
-        manager.stopUpdatingLocation()
-
-        DispatchQueue.main.async {
-            self.lastLocation = location
-            if !self.positionOverridden {
-                self.position = .region(MKCoordinateRegion(
-                    center: location.coordinate,
-                    span: MKCoordinateSpan(latitudeDelta: 0.05, longitudeDelta: 0.05)
-                ))
-            }
-        }
-    }
-    func centerOnUser() {
-        guard let location = lastLocation else { return }
-
-        withAnimation(.easeInOut(duration: 0.6)) {
-            self.position = .region(
-                MKCoordinateRegion(
-                    center: location.coordinate,
-                    span: MKCoordinateSpan(latitudeDelta: 0.08, longitudeDelta: 0.08)
-                )
-            )
-        }
-    }
-}
 
 struct Karte: View {
 
     @StateObject private var locationManager = LocationManager()
-    @State private var mapViewModel = MapViewModel()
+    var mapViewModel: MapViewModel
     @State private var showFilterSheet = false
     @State private var selectedLocationId: UUID?
     @State private var showDetail = false
-    
+
     var userId: UUID?
     @Binding var focusLocation: Location?
 
@@ -70,7 +20,7 @@ struct Karte: View {
                     Annotation(location.name, coordinate: CLLocationCoordinate2D(
                         latitude: location.latitude,
                         longitude: location.longitude
-                            
+
                     )) {
                         LocationPin(imageUrl: location.imageUrls.first)
                             .tag(location.id)
@@ -98,17 +48,12 @@ struct Karte: View {
 
             VStack {
                 Spacer()
-                
+
                 HStack {
                     Spacer()
-                  
+
                     VStack(spacing: 12) {
 
-                        
-                        
-                        
-                        
-                        
                         Button {
                             showFilterSheet = true
                         } label: {
@@ -140,7 +85,7 @@ struct Karte: View {
                                 )
                                 .shadow(color: .black.opacity(0.25), radius: 8, y: 4)
                         }
-                        
+
                         Button {
                             locationManager.centerOnUser()
                         } label: {
@@ -172,56 +117,13 @@ struct Karte: View {
                                 )
                                 .shadow(color: .black.opacity(0.25), radius: 8, y: 4)
                         }
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        /*
-                         Button(action: {
-                             
-                         }) {
-                             Image(systemName: mapViewModel.filter.isEmpty ? "line.3.horizontal.decrease.circle" : "line.3.horizontal.decrease.circle.fill")
-                                 .font(.title2)
-                                 .foregroundColor(mapViewModel.filter.isEmpty ? Color.indigo : .white)
-                                 .frame(width: 50, height: 50)
-                                 .background(mapViewModel.filter.isEmpty ? Color.white : Color.indigo)
-                                 .clipShape(Circle())
-                                 .shadow(radius: 5)
-                                 .fontWeight(.semibold)
-                         }
 
-                         Button(action: {
-                             
-                         }) {
-                             Image(systemName: "location.fill")
-                                 .font(.title2)
-                                 .foregroundColor(Color.indigo)
-                                 .frame(width: 50, height: 50)
-                                 .background(Color.white)
-                                 .clipShape(Circle())
-                                 .shadow(radius: 5)
-                                 .fontWeight(.semibold)
-                         }
-                         */
-
-                        
                     }.padding()
-                    
+
                 }
             }
         }
-     
+
         .sheet(isPresented: $showFilterSheet) {
             LocationFilterView(
                 mapViewModel: mapViewModel,
@@ -257,74 +159,3 @@ struct Karte: View {
         }
     }
 }
-
-// Autocomplete Service
-class SearchCompleter: NSObject, ObservableObject, MKLocalSearchCompleterDelegate {
-
-    @Published var results: [SearchResult] = []
-    var userLocation: CLLocation?
-
-    private let completer = MKLocalSearchCompleter()
-
-    struct SearchResult: Identifiable, Hashable {
-        let id = UUID()
-        let completion: MKLocalSearchCompletion
-        var distance: CLLocationDistance?
-
-        var formattedDistance: String? {
-            guard let distance else { return nil }
-            if distance < 1000 {
-                return "\(Int(distance)) m"
-            } else {
-                return String(format: "%.1f km", distance / 1000)
-            }
-        }
-
-        static func == (lhs: SearchResult, rhs: SearchResult) -> Bool { lhs.id == rhs.id }
-        func hash(into hasher: inout Hasher) { hasher.combine(id) }
-    }
-
-    override init() {
-        super.init()
-        completer.delegate = self
-        completer.resultTypes = .pointOfInterest
-    }
-
-    func update(query: String) {
-        completer.queryFragment = query
-    }
-
-    func completerDidUpdateResults(_ completer: MKLocalSearchCompleter) {
-        let completions = completer.results
-        guard let userLoc = userLocation else {
-            results = completions.map { SearchResult(completion: $0, distance: nil) }
-            return
-        }
-
-        // Resolve distances
-        let group = DispatchGroup()
-        var searchResults: [SearchResult] = []
-
-        for completion in completions {
-            group.enter()
-            let request = MKLocalSearch.Request(completion: completion)
-            let search = MKLocalSearch(request: request)
-            search.start { response, _ in
-                var dist: CLLocationDistance?
-                if let coord = response?.mapItems.first?.placemark.location {
-                    dist = userLoc.distance(from: coord)
-                }
-                searchResults.append(SearchResult(completion: completion, distance: dist))
-                group.leave()
-            }
-        }
-
-        group.notify(queue: .main) {
-            self.results = searchResults.sorted { ($0.distance ?? .greatestFiniteMagnitude) < ($1.distance ?? .greatestFiniteMagnitude) }
-        }
-    }
-}
-
-// Suche mit Live Vorschlägen
-
-
