+++ Sidequest3/MainView.swift
@@ -12,8 +12,8 @@ struct MainView: View {
     @Bindable var authViewModel: AuthViewModel
     var userId: UUID?
     var currentUserId: UUID?
+    var mapViewModel: MapViewModel
     var onShowOnMap: ((Location) -> Void)?
-    @State private var mapViewModel = MapViewModel()
     @State private var viewModel = FeedViewModel()
     @State private var selectedLocation: Location?
     @State private var scrolledId: UUID?
@@ -389,337 +389,5 @@ struct MainView: View {
                 RoundedRectangle(cornerRadius: 28, style: .continuous)
                     .strokeBorder(Color.white.opacity(0.08), lineWidth: 0.8)
             }
-    }
-}
-
-
-
-struct SettingsView: View {
-    
-    @Bindable var authViewModel: AuthViewModel
-    @State private var friendsViewModel = FriendsViewModel()
-    @State private var mapViewModel = MapViewModel()
-    @State private var showLogoutAlert = false
-    @State private var showEditProfile = false
-    @State private var showFriends = false
-    @State private var showShareCard = false
-    @State private var selectedLocation: Location?
-    @State private var pendingCount = 0
-    @Environment(\.dismiss) var dismiss
-
-    var body: some View {
-        if let user = authViewModel.currentUser {
-            NavigationStack {
-                ZStack {
-                    Color(UIColor.systemGray6).ignoresSafeArea()
-                    
-                    VStack(spacing: 20) {
-                        
-                        
-                        
-                        
-                        ScrollView {
-                            Spacer(minLength: 32)
-                            VStack(spacing: 18) {
-                                NavigationLink {
-                                    EditProfileView(authViewModel: authViewModel)
-                                } label: {
-                                    HStack(spacing: 16) {
-                                        
-                                        
-                                        Group {
-                                            if let urlString = user.profileImageUrl,
-                                               let url = URL(string: urlString) {
-                                                AsyncImage(url: url) { image in
-                                                    image
-                                                        .resizable()
-                                                        .scaledToFill()
-                                                        .frame(width: 50, height: 50)
-                                                        .clipShape(Circle())
-                                                } placeholder: {
-                                                    Image(systemName: "person.crop.circle.fill")
-                                                        .resizable()
-                                                        .frame(width: 50, height: 50)
-                                                        .foregroundColor(.indigo)
-                                                }
-                                            } else {
-                                                Image(systemName: "person.crop.circle.fill")
-                                                    .resizable()
-                                                    .frame(width: 50, height: 50)
-                                                    .foregroundColor(.indigo)
-                                            }
-                                        }
-                                        
-                                        VStack(alignment: .leading, spacing: 4) {
-                                            
-                                            Text(user.displayName)
-                                                .foregroundColor(.white)
-                                                .font(.headline)
-                                            
-                                            Text("@\(user.username)")
-                                                .foregroundColor(.gray)
-                                                .font(.subheadline)
-                                        }
-                                        
-                                        Spacer()
-                                    }
-                                    .padding()
-                                    .background(Color(UIColor.systemGray).opacity(0.2))
-                                    .clipShape(RoundedRectangle(cornerRadius: 30))
-                                    .padding(.horizontal)
-
-                                }
-                                
-                                Text("Du bist mit deinem Apple Account angemeldet. Du kannst deinen Namen und dein Profilbild ändern, wenn du auf dein Profil klickst.")
-                                    .font(.footnote)
-                                    .foregroundColor(.gray)
-                                    .padding(.horizontal)
-                                
-                                
-                              
-                                VStack(spacing: 12) {
-                                    NavigationLink {
-                                        NotificationSettingsView(authViewModel: authViewModel)
-                                    } label: {
-                                        SettingsRow(title: "Mitteilungen", value: "Aus")
-                                    }
-                                    NavigationLink {
-                                        LocationSettingsView()
-                                    } label: {
-                                        SettingsRow(title: "Standort", value: "Aus")
-                                    }
-                                    NavigationLink {
-                                        CameraSettingsView()
-                                    } label: {
-                                        SettingsRow(title: "Fotos & Kamera", value: "Aus")
-                                    }
-                                }
-                                .padding(.horizontal)
-                                
-                                
-                                
-                                
-                                
-                                
-                                
-                                
-                                
-                                VStack(alignment: .leading, spacing: 8) {
-                                    
-                                    Text("Datenschutz & Informationen")
-                                        .foregroundColor(.gray)
-                                        .font(.footnote)
-                                        .padding(.horizontal)
-                                    
-                                    VStack(spacing: 0) {
-                                        
-                                        LinkRow(title: "Datenschutz")
-                                        
-                                        Divider()
-                                            .background(Color.gray.opacity(0.3))
-                                        
-                                        LinkRow(title: "Impressum")
-                                        
-                                        Divider()
-                                            .background(Color.gray.opacity(0.3))
-                                        
-                                        LinkRow(title: "Hilfe")
-                                        
-                                    }
-                                    .background(Color(UIColor.systemGray).opacity(0.2))
-                                    .clipShape(RoundedRectangle(cornerRadius: 30))
-                                    .padding(.horizontal)
-                                    
-                                }
-                                
-                                
-                                
-                                
-                                
-                                
-                               
-                                    VStack(alignment: .leading, spacing: 8) {
-                                        Button {
-                                            showLogoutAlert = true
-                                        } label: {
-                                            VStack{
-                                                HStack {
-                                                    Image(systemName: "rectangle.portrait.and.arrow.right")
-                                                        .foregroundColor(.red)
-                                                        .font(.system(size: 14))
-                                                    Text("Abmelden")
-                                                        .foregroundColor(.red)
-                                                    
-                                                    Spacer()
-                                                    
-                                                    
-                                                    
-                                                    
-                                                }
-                                                .padding()
-                                                .background(Color(UIColor.systemGray).opacity(0.2))
-                                                .clipShape(RoundedRectangle(cornerRadius: 30))
-                                            }
-                                            .padding(.horizontal)
-                                        }
-                                    }
-                                    
-                                    .alert("Abmelden?", isPresented: $showLogoutAlert) {
-                                        Button("Abmelden", role: .destructive) {
-                                            authViewModel.signOut()
-                                        }
-                                        Button("Abbrechen", role: .cancel) {}
-                                    } message: {
-                                        Text("Moechtest du dich wirklich abmelden?")
-                                    }
-                                
-
-                                
-                                
-                                HStack(spacing: 4) {
-                                    Image(systemName: "map.fill")
-                                        .foregroundColor(.gray)
-                                        .font(.footnote)
-                                    Text("sidequest")
-                                        .foregroundColor(.gray)
-                                        .font(.footnote)
-                                    
-                                    
-                                }
-                                .padding(.top, 10)
-                                
-                            }
-                        }
-                        .navigationTitle("Einstellungen")
-                        .navigationBarTitleDisplayMode(.inline)
-                        .toolbar {
-                            ToolbarItem(placement: .topBarTrailing) {
-                                Button {
-                                    dismiss()
-                                } label: {
-                                    Image(systemName: "checkmark")
-                                    
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-struct LocationSettingsView: View {
-    
-    var body: some View {
-        Text("Standort Einstellungen")
-            .navigationTitle("Standort")
-    }
-}
-
-struct CameraSettingsView: View {
-    
-    var body: some View {
-        Text("Foto & Kamera Einstellungen")
-            .navigationTitle("Foto & Kamera")
-    }
-}
-
-struct SettingsRow: View {
-    
-    let title: String
-    let value: String
-    
-    var body: some View {
-        HStack {
-            
-            Text(title)
-                .foregroundColor(.white)
-            
-            Spacer()
-            
-            //Text(value)
-            //    .foregroundColor(.gray)
-            
-            Image(systemName: "chevron.right")
-                .foregroundColor(.gray)
-                .font(.system(size: 14))
-        }
-        .padding()
-        .background(Color(UIColor.systemGray).opacity(0.2))
-        .clipShape(RoundedRectangle(cornerRadius: 30))
-    }
-}
-
-struct LinkRow: View {
-    
-    let title: String
-    
-    var body: some View {
-        HStack {
-            Text(title)
-                .foregroundColor(.blue)
-            
-            Spacer()
-        }
-        .padding()
-    }
-}
-
-
-
-
-
-
-struct SkeletonCarouselCard: View {
-    @State private var shimmer = false
-
-    var body: some View {
-        ZStack(alignment: .bottom) {
-            RoundedRectangle(cornerRadius: 28, style: .continuous)
-                .fill(Color.white.opacity(0.06))
-
-            // Centered bottom content matching new card layout
-            VStack(spacing: 8) {
-                // Avatar placeholder
-                Circle()
-                    .fill(Color.white.opacity(0.08))
-                    .frame(width: 40, height: 40)
-
-                // Creator name placeholder
-                RoundedRectangle(cornerRadius: 4)
-                    .fill(Color.white.opacity(0.06))
-                    .frame(width: 80, height: 12)
-
-                // Title placeholder
-                RoundedRectangle(cornerRadius: 4)
-                    .fill(Color.white.opacity(0.10))
-                    .frame(width: 200, height: 26)
-
-                // Address placeholder
-                RoundedRectangle(cornerRadius: 4)
-                    .fill(Color.white.opacity(0.06))
-                    .frame(width: 160, height: 14)
-            }
-            .frame(maxWidth: .infinity)
-            .padding(24)
-        }
-        // Category badge placeholder (top-left)
-        .overlay(alignment: .topLeading) {
-            RoundedRectangle(cornerRadius: 10)
-                .fill(Color.white.opacity(0.08))
-                .frame(width: 60, height: 24)
-                .padding(16)
-        }
-        .clipShape(RoundedRectangle(cornerRadius: 28, style: .continuous))
-        .overlay {
-            RoundedRectangle(cornerRadius: 28, style: .continuous)
-                .strokeBorder(Color.white.opacity(0.08), lineWidth: 0.8)
-        }
-        .shadow(color: .black.opacity(0.32), radius: 32, y: 18)
-        .shadow(color: .black.opacity(0.12), radius: 6, y: 3)
-        .opacity(shimmer ? 0.4 : 1.0)
-        .animation(.easeInOut(duration: 1.0).repeatForever(autoreverses: true), value: shimmer)
-        .onAppear { shimmer = true }
     }
 }
