+++ Sidequest3/Utilities/LocationManager.swift
@@ -1,0 +1,57 @@
+//
+//  LocationManager.swift
+//  Sidequest3
+//
+
+import SwiftUI
+import MapKit
+import CoreLocation
+import Combine
+
+class LocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
+
+    private let manager = CLLocationManager()
+    private var hasSetInitialPosition = false
+    var positionOverridden = false
+    @Published var lastLocation: CLLocation?
+    @Published var position: MapCameraPosition = .region(MKCoordinateRegion(
+        center: CLLocationCoordinate2D(latitude: 52.5200, longitude: 13.4050),
+        span: MKCoordinateSpan(latitudeDelta: 0.05, longitudeDelta: 0.05)
+    ))
+
+    override init() {
+        super.init()
+        manager.delegate = self
+        manager.desiredAccuracy = kCLLocationAccuracyBest
+        manager.requestWhenInUseAuthorization()
+        manager.startUpdatingLocation()
+    }
+
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        guard let location = locations.first, !hasSetInitialPosition else { return }
+        hasSetInitialPosition = true
+        manager.stopUpdatingLocation()
+
+        DispatchQueue.main.async {
+            self.lastLocation = location
+            if !self.positionOverridden {
+                self.position = .region(MKCoordinateRegion(
+                    center: location.coordinate,
+                    span: MKCoordinateSpan(latitudeDelta: 0.05, longitudeDelta: 0.05)
+                ))
+            }
+        }
+    }
+    func centerOnUser() {
+        guard let location = lastLocation else { return }
+
+        withAnimation(.easeInOut(duration: 0.6)) {
+            self.position = .region(
+                MKCoordinateRegion(
+                    center: location.coordinate,
+                    span: MKCoordinateSpan(latitudeDelta: 0.08, longitudeDelta: 0.08)
+                )
+            )
+        }
+    }
+}
