+++ Sidequest3/Views/SkeletonCarouselCard.swift
@@ -1,0 +1,59 @@
+//
+//  SkeletonCarouselCard.swift
+//  Sidequest3
+//
+
+import SwiftUI
+
+struct SkeletonCarouselCard: View {
+    @State private var shimmer = false
+
+    var body: some View {
+        ZStack(alignment: .bottom) {
+            RoundedRectangle(cornerRadius: 28, style: .continuous)
+                .fill(Color.white.opacity(0.06))
+
+            // Centered bottom content matching new card layout
+            VStack(spacing: 8) {
+                // Avatar placeholder
+                Circle()
+                    .fill(Color.white.opacity(0.08))
+                    .frame(width: 40, height: 40)
+
+                // Creator name placeholder
+                RoundedRectangle(cornerRadius: 4)
+                    .fill(Color.white.opacity(0.06))
+                    .frame(width: 80, height: 12)
+
+                // Title placeholder
+                RoundedRectangle(cornerRadius: 4)
+                    .fill(Color.white.opacity(0.10))
+                    .frame(width: 200, height: 26)
+
+                // Address placeholder
+                RoundedRectangle(cornerRadius: 4)
+                    .fill(Color.white.opacity(0.06))
+                    .frame(width: 160, height: 14)
+            }
+            .frame(maxWidth: .infinity)
+            .padding(24)
+        }
+        // Category badge placeholder (top-left)
+        .overlay(alignment: .topLeading) {
+            RoundedRectangle(cornerRadius: 10)
+                .fill(Color.white.opacity(0.08))
+                .frame(width: 60, height: 24)
+                .padding(16)
+        }
+        .clipShape(RoundedRectangle(cornerRadius: 28, style: .continuous))
+        .overlay {
+            RoundedRectangle(cornerRadius: 28, style: .continuous)
+                .strokeBorder(Color.white.opacity(0.08), lineWidth: 0.8)
+        }
+        .shadow(color: .black.opacity(0.32), radius: 32, y: 18)
+        .shadow(color: .black.opacity(0.12), radius: 6, y: 3)
+        .opacity(shimmer ? 0.4 : 1.0)
+        .animation(.easeInOut(duration: 1.0).repeatForever(autoreverses: true), value: shimmer)
+        .onAppear { shimmer = true }
+    }
+}
