+++ backend/package.json
@@ -19,7 +19,8 @@
     "bcryptjs": "^3.0.3",
     "dotenv": "^17.3.1",
     "jsonwebtoken": "^9.0.3",
-    "pg": "^8.20.0"
+    "pg": "^8.20.0",
+    "sharp": "^0.33.0"
   },
   "devDependencies": {
     "jest": "^30.3.0"
